requirements.txt
Pillow>=10.0.0
numpy>=1.19.0